# OPENAPI.yaml
openapi: 3.0.0
info:
  title: Book Store API
  description: A comprehensive REST API for managing an online bookstore with books, authors, customers, orders, and sales reports.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /customers:
    get:
      summary: List all customers
      tags:
        - Customers
      description: Retrieve a list of all customers (protected)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new customer
      tags:
        - Customers
      description: Create a new customer account (protected)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /customers/{customerId}:
    get:
      summary: Get customer by ID
      tags:
        - Customers
      description: Retrieve a specific customer by their ID (protected)
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          description: The ID of the customer to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a customer
      tags:
        - Customers
      description: Update customer information (protected)
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          description: The ID of the customer to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a customer
      tags:
        - Customers
      description: Delete a customer by ID (protected)
      security:
        - BearerAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          description: The ID of the customer to delete
          schema:
            type: integer
      responses:
        '200':
          description: Customer deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
    Customer:
      type: object
      required:
        - name
        - email
        - password
      properties:
        id:
          type: integer
          description: Unique customer identifier
        name:
          type: string
          description: Customer's full name
        email:
          type: string
          format: email
          description: Customer's email address
        password:
          type: string
          format: password
          description: Customer's password (hashed in storage)
        address:
          $ref: '#/components/schemas/Address'
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - JWT token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /customers/register or /customers/login